pyproject.toml
[tool.poetry]
name = "django-cicd"
version = "0.1.0"
description = ""
authors = ["Joel <2016014076@unfv.edu.pe>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
pytest = "^8.4.1"
djangorestframework = "^3.16.0"
pytest-django = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
